+++ CHANGELOG.md
@@ -130,6 +130,7 @@
 - Enh #131: Add `Path` class and corresponding tests for SVG `<path>` element functionality (@terabytesoftw)
 - Bug #132: Update related tests classes to ensure immutability and add missing tests (@terabytesoftw)
 - Enh #133: Add `HasPatternContentUnits` trait and corresponding tests for managing SVG `patternContentUnits` attribute (@terabytesoftw)
+- Enh #134: Add `HasPatternTransform` trait and corresponding tests for managing SVG `patternTransform` attribute (@terabytesoftw)
 
 ## 0.2.0 March 31, 2024